moved serializers to com.ligadata.kamanja.serializer
