+++ trunk/SampleApplication/HelloWorld/config/HelloWorld_Adapter_Binding.json
@@ -11,7 +11,6 @@
   {
     "AdapterName": "testout_1",
     "MessageNames": [
-      "com.ligadata.kamanja.samples.messages.inmsg1",
       "com.ligadata.kamanja.samples.messages.msg1",
       "com.ligadata.kamanja.samples.messages.outmsg1"
     ],
