+++ trunk/Utils/AdapterSerializers/KBinarySerDeser/src/main/scala/com/ligadata/kamanja/serializer/KBinarySerDeser.scala
@@ -1,4 +1,4 @@
-package org.kamanja.serdeser.kbinary
+package com.ligadata.kamanja.serializer
 
 import com.ligadata.BaseTypes._
